requirements.txt
python-dotenv
transformers>=4.30.0
huggingface_hub>=0.16.0
torch>=2.0.0
accelerate
bitsandbytes
